chore: Renaming and documenting
The expression is probably not just a parser anymore and the README does not say that constraints have been added.
